chatBot/requirements.txt
Django
djangorestframework
requests
python-dotenv
groq